requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
scikit-learn==1.2.2
pandas==2.1.3
